design: 스켈레톤 컴포넌트 생성
