+++ apps/web/src/components/midpoint-result/style.css.ts
@@ -1,6 +1,5 @@
 import { zIndex } from "@repo/z-index";
 import { style } from "@vanilla-extract/css";
-import { keyframes } from "@vanilla-extract/css";
 
 export const mapContainer = style({
   display: "flex",
@@ -42,43 +41,3 @@ export const headerSkeletonWrapper = style({
   justifyContent: "center",
   width: "100%",
 });
-
-const skeletonKeyframes = keyframes({
-  from: {
-    backgroundPosition: "200% 0",
-  },
-  to: {
-    backgroundPosition: "-200% 0",
-  },
-});
-
-export const skeletonAnimation = style({
-  animationName: skeletonKeyframes,
-  animationDuration: "1.5s",
-  animationIterationCount: "infinite",
-  animationTimingFunction: "linear",
-});
-
-export const roomNameSkeletonStyle = style([
-  skeletonAnimation,
-  {
-    background: "linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%)",
-    backgroundSize: "200% 100%",
-    width: "120px",
-    height: "24px",
-    borderRadius: "4px",
-    margin: "0 auto",
-  },
-]);
-
-export const participantsSkeletonStyle = style([
-  skeletonAnimation,
-  {
-    background: "linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%)",
-    backgroundSize: "200% 100%",
-    width: "32px",
-    height: "24px",
-    borderRadius: "4px",
-    display: "inline-block",
-  },
-]);
